pyproject.toml
[tool.poetry]
name = "qa-bot"
version = "0.1.0"
description = ""
authors = ["HypedAugust <44738421+HypedAugust@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1.0.1"
langchain = "^0.2.11"
langchain-openai = "^0.1.17"
gradio = "^4.39.0"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
